backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pydantic = "^2.9.2"
requests = "^2.32.3"
flask-swagger-ui = "^4.11.1"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"




[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
